Correctly check if the node is using system path
